Break dashboard pages out into separate always-rendered, conditionally-hidden components
