+++ assets/css/dashboard.scss
@@ -1,16 +1,1 @@
-.place-list__header-row {
-  display: flex;
-  align-items: center;
-  margin-bottom: 24px;
 
-  .place-list__sort-label {
-    color: white;
-    font-weight: 900;
-    font-size: 16px;
-    line-height: 24px;
-    letter-spacing: 1px;
-    margin-right: auto;
-    cursor: pointer;
-    width: fit-content;
-  }
-}
