+++ assets/css/dashboard/alerts-page.scss
@@ -1,0 +1,3 @@
+.alerts-page.alerts-page--hidden {
+  display: none;
+}
