+++ assets/css/dashboard/overrides-page.scss
@@ -1,0 +1,3 @@
+.overrides-page.overrides-page--hidden {
+  display: none;
+}
