+++ assets/css/dashboard/places-page.scss
@@ -1,0 +1,20 @@
+.places-page.places-page--hidden {
+  display: none;
+}
+
+.place-list__header-row {
+  display: flex;
+  align-items: center;
+  margin-bottom: 24px;
+
+  .place-list__sort-label {
+    color: white;
+    font-weight: 900;
+    font-size: 16px;
+    line-height: 24px;
+    letter-spacing: 1px;
+    margin-right: auto;
+    cursor: pointer;
+    width: fit-content;
+  }
+}
