+++ assets/css/screenplay.scss
@@ -1,6 +1,9 @@
 @import "~bootstrap/scss/bootstrap";
 @import "colors.scss";
 @import "dashboard.scss";
+@import "dashboard/places-page.scss";
+@import "dashboard/alerts-page.scss";
+@import "dashboard/overrides-page.scss";
 @import "place-row.scss";
 @import "places-action-bar.scss";
 @import "filter-dropdown.scss";
